Resample and Correlation Plot
This is resampling and correlation code, which runs in R Studio. This code helps to make the variable present in the same extension, sample size, no. of rows and columns, which is useful for suitability modelling. Hence, it makes all the varibales to be present in a desired format. It also helps to read multiple no. of variables and export in any desired format.